+++ src/Shared/SharedConstants.swift
@@ -1,0 +1,10 @@
+//
+//  SharedConstants.swift
+//  Reed
+//
+
+import Foundation
+
+enum SharedConstants {
+    static let appGroup = "group.net.domzilla.reed"
+}